:sparkles: Add casbin component
添加 casbin 组件

+++ server/pkg/utils/jwt/jwt.go
@@ -9,7 +9,7 @@ var  jwtSignKey = []byte("qwerasdf")
 var  TokenExpired = errors.New("token expired")
 
 type CustomClaims struct {
-	ID uint
+	ID uint64
 	UUID string
 	Name string
 	jwt.StandardClaims
